Restructured problem code
