+++ ctsb/problems/__init__.py
@@ -10,12 +10,6 @@ from ctsb.problems.control import ControlProblem
 
 
 problem_register(
-    id='Pendulum-v0',
-    entry_point='ctsb.problems.control.pybullet:InvertedPendulumSwingupBulletEnv'
-)
-
-
-problem_register(
     id='CartPole-v0',
     entry_point='ctsb.problems.control.pybullet:CartPole'
 )
