+++ ctsb/problems/control/pybullet/__init__.py
@@ -1,6 +1,5 @@
 # problems/control/pybullet problem init file
 
-from ctsb.problems.control.pybullet.pendulum import InvertedPendulumSwingupBulletEnv
 from ctsb.problems.control.pybullet.kuka_diverse_object import Kuka_diverse_object
 from ctsb.problems.control.pybullet.pybullet_problem import PyBulletProblem
 from ctsb.problems.control.pybullet.cartpole import CartPole
