+++ ctsb/problems/control/tests/test_kuka_stationary.py
