+++ ctsb/problems/core.py
@@ -28,9 +28,9 @@ class Problem(object):
 
     def __str__(self):
         if self.spec is None:
-            return '<{} instance> call object help() method for info'.format(type(self).__name__)
+            return '<{} instance>'.format(type(self).__name__)
         else:
-            return '<{}<{}>> call object help() method for info'.format(type(self).__name__, self.spec.id)
+            return '<{}<{}>>'.format(type(self).__name__, self.spec.id)
 
     def __enter__(self):
         return self
