+++ ctsb/problems/time_series/tests/test_sp500.py
