+++ ctsb/utils/experiment.py
@@ -127,20 +127,5 @@ class Experiment(object):
     def help(self):
         # prints information about this class and its methods
         raise NotImplementedError
-    '''
-    def __str__(self):
-        if self.spec is None:
-            return '<{} instance> call object help() method for info'.format(type(self).__name__)
-        else:
-            return '<{}<{}>> call object help() method for info'.format(type(self).__name__, self.spec.id)
-    
-    def __enter__(self):
-        return self
-
-    def __exit__(self, *args):
-        self.close()
-        # propagate exception
-        return False
-    '''
